[WFCORE-2691] realm identity resources replaced by realm operations
