+++ elytron/src/main/java/org/wildfly/extension/elytron/_private/ElytronSubsystemMessages.java
@@ -330,7 +330,7 @@ public interface ElytronSubsystemMessages extends BasicLogger {
     @Message(id = 917, value = "Password cannot be resolved for dir-context")
     StartException dirContextPasswordCannotBeResolved(@Cause Exception cause);
 
-    @Message(id = 918, value = "Invalid user name '%s' because the realm %s only supports lower case alias names")
+    @Message(id = 918, value = "Invalid user name '%s' because the realm %s only supports lower case user names")
     OperationFailedException invalidUsername(String username, String realmName);
 
     @Message(id = 919, value = "Invalid alias name '%s' because the credential store %s only supports lower case alias names")
@@ -343,37 +343,37 @@ public interface ElytronSubsystemMessages extends BasicLogger {
     OperationFailedException identityAlreadyExists(final String principalName);
 
     @Message(id = 1001, value = "Could not create identity with name [%s].")
-    OperationFailedException couldNotCreateIdentity(final String principalName, @Cause Exception cause);
+    RuntimeException couldNotCreateIdentity(final String principalName, @Cause Exception cause);
 
     @Message(id = 1002, value = "Identity with name [%s] not found.")
     String identityNotFound(final String principalName);
 
     @Message(id = 1003, value = "Could not delete identity with name [%s].")
-    OperationFailedException couldNotDeleteIdentity(final String principalName, @Cause Exception cause);
+    RuntimeException couldNotDeleteIdentity(final String principalName, @Cause Exception cause);
 
     @Message(id = 1004, value = "Identity with name [%s] not authorized.")
     String identityNotAuthorized(final String principalName);
 
     @Message(id = 1005, value = "Could not read identity [%s] from security domain [%s].")
-    OperationFailedException couldNotReadIdentity(final String principalName, final ServiceName domainServiceName, @Cause Exception cause);
+    RuntimeException couldNotReadIdentity(final String principalName, final ServiceName domainServiceName, @Cause Exception cause);
 
     @Message(id = 1006, value = "Unsupported password type [%s].")
     RuntimeException unsupportedPasswordType(final Class passwordType);
 
     @Message(id = 1007, value = "Could not read identity with name [%s].")
-    OperationFailedException couldNotReadIdentity(final String principalName, @Cause Exception cause);
+    RuntimeException couldNotReadIdentity(final String principalName, @Cause Exception cause);
 
     @Message(id = 1008, value = "Failed to obtain the authorization identity.")
-    OperationFailedException couldNotObtainAuthorizationIdentity(@Cause Exception cause);
+    RuntimeException couldNotObtainAuthorizationIdentity(@Cause Exception cause);
 
     @Message(id = 1009, value = "Failed to add attribute.")
-    OperationFailedException couldNotAddAttribute(@Cause Exception cause);
+    RuntimeException couldNotAddAttribute(@Cause Exception cause);
 
     @Message(id = 1010, value = "Failed to remove attribute.")
-    OperationFailedException couldNotRemoveAttribute(@Cause Exception cause);
+    RuntimeException couldNotRemoveAttribute(@Cause Exception cause);
 
     @Message(id = 1011, value = "Could not create password.")
-    OperationFailedException couldNotCreatePassword(@Cause Exception cause);
+    RuntimeException couldNotCreatePassword(@Cause Exception cause);
 
     @Message(id = 1012, value = "Unexpected password type [%s].")
     OperationFailedException unexpectedPasswordType(final String passwordType);
@@ -388,7 +388,7 @@ public interface ElytronSubsystemMessages extends BasicLogger {
     IllegalStateException unableToPerformOutflow(String identityName, @Cause Exception cause);
 
     @Message(id = 1016, value = "Server '%s' not known")
-    OperationFailedException serverNotKnown(final String serverAddedd, @Cause UnknownHostException e);
+    OperationFailedException serverNotKnown(final String server, @Cause UnknownHostException e);
 
     @Message(id = 1017, value = "Invalid value for cipher-suite-filter. %s")
     OperationFailedException invalidCipherSuiteFilter(@Cause Throwable cause, String causeMessage);
